Refactor enchantment printing
