+++ runtime/locale/en/enchantment.hcl
@@ -1,0 +1,81 @@
+locale {
+    enchantment {
+        level = "#"
+
+        with_parameters {
+            attribute {
+                in_food {
+                    increases =
+                    decreases =
+                }
+                other {
+                    increases =
+                    decreases =
+                }
+            }
+
+            resistance {
+                increases =
+                decreases =
+            }
+
+            skill {
+                increases =
+                decreases =
+            }
+
+            skill_maintenance {
+                in_food =
+                other =
+            }
+
+            extra_damage =
+            invokes =
+            can_be_loaded_with =
+        }
+
+        no_parameters {
+            _21 = "causes random teleport."
+            _22 = "prevents you from teleporting."
+            _23 = "negates the effect of blindness."
+            _24 = "negates the effect of paralysis."
+            _25 = "negates the effect of confusion."
+            _26 = "negates the effect of fear."
+            _27 = "negates the effect of sleep."
+            _28 = "negates the effect of poison."
+            _29 = "speeds up your travel progress."
+            _30 = "protects you from Etherwind."
+            _31 = "negates the effect of being stranded by bad weather."
+            _32 = "floats you."
+            _33 = "protects you from mutation."
+            _34 = "enhances your spells."
+            _35 = "allows you to see invisible creatures."
+            _36 = "absorbs stamina from an enemy."
+            _37 = "brings an end."
+            _38 = "absorbs MP from an enemy."
+            _39 = "gives you a chance to throw an absolute piercing attack."
+            _40 = "occasionally stops time."
+            _41 = "protects you from thieves."
+            _42 = "allows you to digest rotten food."
+            _43 = "protects you from cursing words."
+            _44 = "increases your chance to deliver critical hits."
+            _45 = "sucks blood of the wielder."
+            _46 = "disturbs your growth."
+            _47 = "attracts monsters."
+            _48 = "prevents aliens from entering your body."
+            _49 = "increases the qualitiy of reward."
+            _50 = "increases the chance of extra melee attack."
+            _51 = "increases the chance of extra ranged attack."
+            _52 = "decreases physical damage you take."
+            _53 = "sometimes nullifies damage you take."
+            _54 = "deals cut damage to the attacker."
+            _55 = "diminishs bleeding."
+            _56 = "catches signals from God."
+            _57 = "inflicts massive damage to dragons."
+            _58 = "inflicts massive damage to undeads."
+            _59 = "reveals religion."
+            _60 = "makes audience drunk with haunting tones."
+            _61 = "inflicts massive damage to Gods."
+        }
+    }
+}
